26x/fix/pre comp change (#92)
* Added Skidpad

* Idk how if stopped compiling

* Auto-generate symbol files [skip ci]

* Added Skidpad

* Idk how if stopped compiling

* Updated Optimal Yaw Rate

* Revert "Updated Optimal Yaw Rate"

This reverts commit 73a0826234ffd335d3ad6cd200b9994a736e8618.

* mc_power fix

* nit

* Nit

* Fixing DIM Screen and DV State Machine

* Fixed DIM Screen

* RR Fixes

* Made FDCAN More supportive of 32 bit items

* Adding In the CubeMars Filter

* Only commenting out cubemars

* Added VSM_HEART_BEAT Forwarding

* Changed CAN-ID For CubeMars

* update reqVSM glv on request

* Only check for tank pressure when ASMS is on

* Added Check to ensure tanks are empty on state up

* Fix weird VSM bugs for DV_errors

* Added in compute and cubemars checks

* Fixed HVC Compiling

* Updated TSSI Code To Latch

* Updated TSSI to not be time based

* nit

* Trying smth

* Loosening DV_Error_requirments

* add in brakelight

* allow for new heartbeat CAN stuff in symbol file

* Auto-generate symbol files [skip ci]

* fix stategetVSM function call

* remove racing screen & add back config screen to dim state machine

* Fixing Some Things

* Sending BSPD Current

* Controls on

* Fixing Controls Off

* Auto-generate symbol files [skip ci]

* Made HV_Current not shit

* Auto-generate symbol files [skip ci]

* added tuning parameters

* Forward AS_finished

* Reverted Bad Merge

* Auto-generate symbol files [skip ci]

* Stupid VS Code Autocorrect

* Changed Negative Torque Convention

* Modified transfer function for VSM_HE

* Auto-generate symbol files [skip ci]

* Added Comments

* Revert "Changed Negative Torque Convention"

This reverts commit 449c88ca5eea146296916d267ffb2ab56c43a089.

* Make negative torques negative

* Adding Regen

* Fixed build issue

* build

* Updating DV Sign Convention

* remove rss check from glv

* Fixing Previous Commit

* Code Clean Up

* AS_Finished_Bug

* Added case that if sensoric times out we head to fast mode

* Move Forwarding 10Hz

* DV state machine fixes & more error handling

* Auto-generate symbol files [skip ci]

---------

+++ stm32f413-drivers/CMR/include/CMR/fdcan.h
@@ -70,7 +70,7 @@ int cmr_canRXMetaTimeoutError(const cmr_canRXMeta_t *meta, TickType_t now_ms);
 typedef struct cmr_can cmr_can_t;
 
 typedef void (*cmr_canRXCallback_t)(
-    cmr_can_t *can, uint16_t canID, const void *data, size_t dataLen
+    cmr_can_t *can, uint32_t canID, const void *data, size_t dataLen
 );
